+++ core/execution/spot/market.go
@@ -819,26 +819,29 @@ func (m *Market) getNewPeggedPrice(order *types.Order) (*num.Uint, error) {
 		return num.UintZero(), common.ErrUnableToReprice
 	}
 
-	offset, _ := num.UintFromDecimal(order.PeggedOrder.Offset.ToDecimal().Mul(m.priceFactor))
+	// we're converting both offset and tick size to asset decimals so we can adjust the price (in asset) directly
+	priceInMarket, _ := num.UintFromDecimal(price.ToDecimal().Div(m.priceFactor))
 	if order.Side == types.SideSell {
-		price = price.AddSum(offset)
+		priceInMarket.AddSum(order.PeggedOrder.Offset)
 		// this can only happen when pegged to mid, in which case we want to round to the nearest *better* tick size
 		// but this can never cross the mid by construction as the the minimum offset is 1 tick size and all prices must be
 		// whole multiples of tick size.
-		if mod := num.UintZero().Mod(price, m.mkt.TickSize); !mod.IsZero() {
-			price.Sub(price, mod)
+		if mod := num.UintZero().Mod(priceInMarket, m.mkt.TickSize); !mod.IsZero() {
+			priceInMarket.Sub(priceInMarket, mod)
 		}
+		price, _ := num.UintFromDecimal(priceInMarket.ToDecimal().Mul(m.priceFactor))
 		return price, nil
 	}
 
-	if price.LTE(offset) {
+	if priceInMarket.LTE(order.PeggedOrder.Offset) {
 		return num.UintZero(), common.ErrUnableToReprice
 	}
 
-	price.Sub(price, offset)
-	if mod := num.UintZero().Mod(price, m.mkt.TickSize); !mod.IsZero() {
-		price = num.UintZero().Sub(price.AddSum(m.mkt.TickSize), mod)
+	priceInMarket.Sub(priceInMarket, order.PeggedOrder.Offset)
+	if mod := num.UintZero().Mod(priceInMarket, m.mkt.TickSize); !mod.IsZero() {
+		priceInMarket = num.UintZero().Sub(priceInMarket.AddSum(m.mkt.TickSize), mod)
 	}
+	price, _ = num.UintFromDecimal(priceInMarket.ToDecimal().Mul(m.priceFactor))
 
 	return price, nil
 }
@@ -971,6 +974,7 @@ func (m *Market) leaveAuction(ctx context.Context, now time.Time) {
 }
 
 // validateOrder checks that the order parameters are valid for the market.
+// NB: price in market, tickSize in market decimals.
 func (m *Market) validateOrder(ctx context.Context, order *types.Order) (err error) {
 	defer func() {
 		if err != nil {